Make WCS.jl file
